refactor(holochain): move is_signature_valid to conductor
